refactor(phase-10): extract large files into focused single-responsibility modules
6 extractions: SortControls toolbar logic (sync/prune/collapse), IssueHeader info content builder,
issue-creation-modal helpers + file-focus, dashboard-issue-action confirmations, IssueManager block-edit,
OverflowPanel layout settings into child component. Zero behavioral changes.
