+++ src/dashboard/dashboard-issue-actions.ts
@@ -1,10 +1,5 @@
 import { Menu, Notice } from 'obsidian';
 import TasksDashboardPlugin from '../../main';
-import { ArchiveConfirmationModal } from '../modals/archive-confirmation-modal';
-import {
-	DeleteConfirmationModal,
-	type DeleteConfirmationResult
-} from '../modals/delete-confirmation-modal';
 import { FolderPathModal } from '../modals/FolderPathModal';
 import { GitHubLinksModal } from '../modals/github-links-modal';
 import { openIssueCreationModal, openPrioritySelectionModal } from '../modals/issue-creation-modal';
@@ -14,9 +9,12 @@ import { getGitHubLinkType, isGitHubWebUrl } from '../utils/github';
 import { formatGitHubLinkLabel } from '../utils/github-helpers';
 import { parseGitHubRepoFullName } from '../utils/github-url';
 import { getIssueFolderStorageKey } from '../issues/issue-manager-shared';
-import { recordDeletedIssueGitHubUrls } from '../issues/issue-manager-settings';
 import type { PlatformService } from '../utils/platform';
 import { isNonEmptyString } from '../utils/string-utils';
+import {
+	handleArchiveWithConfirmation,
+	handleDeleteWithConfirmation
+} from './dashboard-issue-action-confirmations';
 import { openIssueColorDropdown } from './dashboard-issue-color-dropdown';
 import { ControlParams, IssueActionDescriptor } from './dashboard-renderer-types';
 
@@ -386,39 +384,7 @@ export const buildIssueActionDescriptors = (options: {
 				return;
 			}
 
-			void Promise.all([
-				plugin.issueManager.hasAssociatedWorktree(dashboard, params.issue),
-				plugin.progressTracker.getProgress(params.path)
-			])
-				.then(([hasAssociatedWorktree, progress]) => {
-					const unfinishedTaskCount = progress.total - progress.done;
-					if (!hasAssociatedWorktree && unfinishedTaskCount === 0) {
-						void plugin.issueManager.archiveIssue(dashboard, params.issue);
-						return;
-					}
-
-					const modal = new ArchiveConfirmationModal(
-						plugin.app,
-						params.name,
-						hasAssociatedWorktree,
-						unfinishedTaskCount,
-						(result) => {
-							if (!result.confirmed) {
-								return;
-							}
-							if (result.removeWorktree) {
-								void plugin.issueManager.removeWorktree(dashboard, params.issue, {
-									skipScriptConfirmation: true
-								});
-							}
-							void plugin.issueManager.archiveIssue(dashboard, params.issue);
-						}
-					);
-					modal.open();
-				})
-				.catch(() => {
-					new Notice(`Could not archive: ${params.issue}`);
-				});
+			handleArchiveWithConfirmation({ plugin, dashboard, params });
 		}
 	});
 
@@ -430,54 +396,7 @@ export const buildIssueActionDescriptors = (options: {
 		shouldRender: true,
 		faded: false,
 		onClick: () => {
-			void Promise.all([
-				plugin.issueManager.hasAssociatedWorktree(dashboard, params.issue),
-				plugin.progressTracker.getProgress(params.path)
-			])
-				.then(([hasAssociatedWorktree, progress]) => {
-					const unfinishedTaskCount = progress.total - progress.done;
-					const modal = new DeleteConfirmationModal(
-						plugin.app,
-						params.name,
-						hasAssociatedWorktree,
-						unfinishedTaskCount,
-						hasAssociatedWorktree
-							? plugin.settings.deleteIssueRemoveWorktreeByDefault
-							: false,
-						(checked: boolean) => {
-							plugin.settings.deleteIssueRemoveWorktreeByDefault = checked;
-							void plugin.saveSettings();
-						},
-						(result: DeleteConfirmationResult) => {
-							if (!result.confirmed) {
-								return;
-							}
-							if (result.removeWorktree) {
-								void plugin.issueManager.removeWorktree(dashboard, params.issue, {
-									skipScriptConfirmation: true
-								});
-							}
-							void plugin.issueManager
-								.deleteIssue(dashboard, params.issue)
-								.then(() => {
-									const webUrls = params.githubLinks.filter(isGitHubWebUrl);
-									if (
-										recordDeletedIssueGitHubUrls(
-											plugin.settings,
-											dashboard.id,
-											webUrls
-										)
-									) {
-										void plugin.saveSettings();
-									}
-								});
-						}
-					);
-					modal.open();
-				})
-				.catch(() => {
-					new Notice(`Could not delete: ${params.issue}`);
-				});
+			handleDeleteWithConfirmation({ plugin, dashboard, params });
 		}
 	});
