+++ src/modals/issue-creation-modal.ts
@@ -1,12 +1,8 @@
-import { App, MarkdownView, Modal, Notice, TFile } from 'obsidian';
+import { App, Modal, Notice } from 'obsidian';
 import type { Component } from 'svelte';
 import { mount, unmount } from 'svelte';
 import TasksDashboardPlugin from '../../main';
-import { getErrorMessage } from '../settings/settings-helpers';
-import { DashboardConfig, GitHubIssueMetadata, GitHubRepository, Priority } from '../types';
-import { getGitHubLinkType } from '../utils/github';
-import { parseGitHubRepoFullName, parseGitHubUrl } from '../utils/github-url';
-import { findIssueColorPaletteIndex } from '../utils/issue-colors';
+import { DashboardConfig, Priority } from '../types';
 import IssueCreationWizard from '../components/modals/IssueCreationWizard.svelte';
 import PrioritySelector from '../components/modals/PrioritySelector.svelte';
 import ManualGitHubLinkContent from '../components/modals/ManualGitHubLinkContent.svelte';
@@ -19,19 +15,20 @@ import type {
 	QuickCreateDefaults,
 	WorktreeCreationContext
 } from './issue-creation-types';
+import {
+	createIssueWithGitHub,
+	getIssueLinkedRepositoryFromLinks,
+	getPrefilledIssueName
+} from './issue-creation-helpers';
 
-interface CreateIssueRequest {
-	name: string;
-	priority: Priority;
-	dashboard: DashboardConfig;
-	color?: string;
-	worktree?: boolean;
-	worktreeOriginFolder?: string;
-	worktreeScriptDirectory?: string;
-	worktreeBaseRepository?: string;
-	githubLink?: string;
-	githubMetadata?: GitHubIssueMetadata;
-}
+export type { AssignedIssueCreationOptions } from './issue-creation-helpers';
+
+export {
+	createIssueWithGitHub,
+	createIssueWithRepoLink,
+	createIssueWithNotice,
+	getPrefilledIssueName
+} from './issue-creation-helpers';
 
 interface IssueCreationModalOptions {
 	mode?: IssueCreationMode;
@@ -41,67 +38,6 @@ interface IssueCreationModalOptions {
 	githubSelection?: GitHubSelectionContext;
 	worktreeContext?: WorktreeCreationContext;
 	quickCreateDefaults?: QuickCreateDefaults;
-}
-
-function getIssueLinkedRepositoryFromLinks(githubLinks: string[] | undefined): string | undefined {
-	if (githubLinks === undefined || githubLinks.length === 0) {
-		return undefined;
-	}
-
-	for (const link of githubLinks) {
-		if (getGitHubLinkType(link) !== 'repository') {
-			continue;
-		}
-
-		const repository = parseGitHubRepoFullName(link);
-		if (repository !== undefined && repository !== '') {
-			return repository;
-		}
-	}
-
-	for (const link of githubLinks) {
-		const parsed = parseGitHubUrl(link);
-		if (parsed !== undefined) {
-			return `${parsed.owner}/${parsed.repo}`;
-		}
-	}
-
-	return undefined;
-}
-
-async function createIssueWithNotice(
-	app: App,
-	plugin: TasksDashboardPlugin,
-	request: CreateIssueRequest
-): Promise<void> {
-	try {
-		const issue = await plugin.issueManager.createIssue({
-			...request,
-			worktreeColor: request.color
-		});
-		if (request.color !== undefined) {
-			plugin.settings.issueColors[issue.id] = request.color;
-			const paletteIndex = findIssueColorPaletteIndex(request.color);
-			if (paletteIndex >= 0) {
-				plugin.settings.lastUsedColorIndex = paletteIndex;
-			}
-			await plugin.saveSettings();
-		}
-		if (request.worktree === true) {
-			void plugin.issueManager.setupWorktree(
-				request.dashboard,
-				issue.id,
-				request.name,
-				request.color,
-				request.worktreeOriginFolder,
-				request.worktreeScriptDirectory
-			);
-		}
-		new Notice(`Created issue: ${request.name}`);
-		await openFileAndFocusEnd(app, issue.filePath);
-	} catch (error) {
-		new Notice(`Error creating issue: ${getErrorMessage(error)}`);
-	}
 }
 
 class IssueCreationModal extends Modal {
@@ -299,100 +235,15 @@ class ManualGitHubLinkFirstModal extends SvelteModal {
 	}
 }
 
-async function openFileAndFocusEnd(app: App, filePath: string): Promise<void> {
-	const file = app.vault.getAbstractFileByPath(filePath);
-	if (file instanceof TFile) {
-		const leaf = app.workspace.getLeaf();
-		await leaf.openFile(file);
-		window.setTimeout(() => {
-			const view = app.workspace.getActiveViewOfType(MarkdownView);
-			if (view?.editor) {
-				const editor = view.editor;
-				const lastLine = editor.lastLine();
-				const tasksHeadingLine = findTasksHeadingLine(editor, lastLine);
-
-				if (tasksHeadingLine === undefined) {
-					const lastLineLength = editor.getLine(lastLine).length;
-					editor.setCursor({ line: lastLine, ch: lastLineLength });
-					editor.focus();
-					return;
-				}
-
-				const targetLine = getTasksSectionTargetLine(editor, tasksHeadingLine, lastLine);
-				const targetLineLength = editor.getLine(targetLine).length;
-				editor.setCursor({ line: targetLine, ch: targetLineLength });
-				editor.focus();
-			}
-		}, 100);
-	}
-}
-
-function findTasksHeadingLine(
-	editor: MarkdownView['editor'],
-	lastLine: number
-): number | undefined {
-	for (let lineIndex = 0; lineIndex <= lastLine; lineIndex += 1) {
-		const line = editor.getLine(lineIndex).trim();
-		if (line.toLowerCase() === '## tasks') {
-			return lineIndex;
-		}
-	}
-
-	return undefined;
-}
-
-function getTasksSectionTargetLine(
-	editor: MarkdownView['editor'],
-	tasksHeadingLine: number,
-	lastLine: number
-): number {
-	let sectionEndLine = lastLine;
-
-	for (let lineIndex = tasksHeadingLine + 1; lineIndex <= lastLine; lineIndex += 1) {
-		const line = editor.getLine(lineIndex).trim();
-		if (line.startsWith('## ')) {
-			sectionEndLine = lineIndex - 1;
-			break;
-		}
-	}
-
-	while (sectionEndLine > tasksHeadingLine && editor.getLine(sectionEndLine).trim() === '') {
-		sectionEndLine -= 1;
-	}
-
-	if (sectionEndLine <= tasksHeadingLine) {
-		const candidateLine = tasksHeadingLine + 1;
-		return candidateLine <= lastLine ? candidateLine : tasksHeadingLine;
-	}
-
-	return sectionEndLine;
-}
-
-function getPrefilledIssueName(metadata: GitHubIssueMetadata | undefined): string | undefined {
-	if (metadata === undefined) {
-		return undefined;
-	}
-
-	const firstFourWords = metadata.title
-		.trim()
-		.split(/\s+/)
-		.filter((word) => word !== '')
-		.slice(0, 4)
-		.join(' ');
-	return firstFourWords === '' ? `${metadata.number}` : `${metadata.number} ${firstFourWords}`;
-}
-
-export interface AssignedIssueCreationOptions {
-	dashboard: DashboardConfig;
-	githubMetadata: GitHubIssueMetadata;
-	githubUrl: string;
-	quickCreateDefaults?: QuickCreateDefaults;
-}
-
 export const openAssignedIssueNamePrompt = (
 	app: App,
 	plugin: TasksDashboardPlugin,
-	options: AssignedIssueCreationOptions
+	options: {
+		dashboard: DashboardConfig;
+		githubMetadata: import('../types').GitHubIssueMetadata;
+		githubUrl: string;
+		quickCreateDefaults?: QuickCreateDefaults;
+	}
 ): void => {
 	const prefilledName = getPrefilledIssueName(options.githubMetadata);
 	new IssueCreationModal(app, plugin, options.dashboard, {
@@ -406,59 +257,6 @@ export const openAssignedIssueNamePrompt = (
 		quickCreateDefaults: options.quickCreateDefaults
 	}).open();
 };
-
-export async function createIssueWithRepoLink(
-	app: App,
-	plugin: TasksDashboardPlugin,
-	dashboard: DashboardConfig,
-	issueName: string,
-	priority: Priority,
-	repository: GitHubRepository,
-	color?: string,
-	mode: IssueCreationMode = 'standard',
-	worktreeOriginFolder?: string,
-	worktreeBaseRepository?: string
-): Promise<void> {
-	const repoUrl = `https://github.com/${repository.fullName}`;
-	await createIssueWithNotice(app, plugin, {
-		name: issueName,
-		priority,
-		githubLink: repoUrl,
-		color,
-		worktree: mode === 'worktree',
-		worktreeOriginFolder,
-		worktreeBaseRepository,
-		dashboard
-	});
-}
-
-export async function createIssueWithGitHub(
-	app: App,
-	plugin: TasksDashboardPlugin,
-	dashboard: DashboardConfig,
-	issueName: string,
-	priority: Priority,
-	githubUrl?: string,
-	githubMetadata?: GitHubIssueMetadata,
-	color?: string,
-	mode: IssueCreationMode = 'standard',
-	worktreeOriginFolder?: string,
-	worktreeBaseRepository?: string,
-	worktreeScriptDirectory?: string
-): Promise<void> {
-	await createIssueWithNotice(app, plugin, {
-		name: issueName,
-		priority,
-		githubLink: githubUrl,
-		githubMetadata,
-		color,
-		worktree: mode === 'worktree',
-		worktreeOriginFolder,
-		worktreeScriptDirectory,
-		worktreeBaseRepository,
-		dashboard
-	});
-}
 
 export const openWorktreeIssueCreationModal = (
 	app: App,
